docstrings and restructuring/cleaning of the files
