+++ src/esm_tests/cli.py
@@ -8,6 +8,7 @@ from loguru import logger
 
 from .info import *
 from .initialization import *
+from .output import *
 from .read_shipped_data import *
 from .repos import *
 from .test_utilities import *
