+++ src/esm_tests/tests.py
@@ -1,16 +1,14 @@
 import copy
-import difflib
 import glob
 import os
 import re
 import shutil
 import time
-import colorama
 import yaml
 from esm_parser import determine_computer_from_hostname
-from esm_runscripts import color_diff
 from loguru import logger
 
+from .output import *
 from .read_shipped_data import *
 from .test_utilities import *
 
@@ -103,7 +101,7 @@ def comp_test(info):
 
     Parameters
     ----------
-    info : dict
+    info : esm_tests.Info
         Dictionary that contains the testing info.
     """
     # Load dictionaries from ``info``
@@ -274,7 +272,7 @@ def run_test(info):
 
     Parameters
     ----------
-    info : dict
+    info : esm_tests.Info
         Dictionary that contains the testing info.
     """
     # Load dictionaries from ``info``
@@ -452,7 +450,7 @@ def check(info, mode, model, version, out, script, v):
 
     Parameters
     ----------
-    info : dict
+    info : esm_tests.Info
         Dictionary that contains the testing info.
     mode : str
         Mode of the check.
@@ -619,7 +617,7 @@ def experiment_state_action(info, message, no_err, v, finished_runs, cc, subc, m
 
     Parameters
     ----------
-    info : dict
+    info : esm_tests.Info
         Dictionary that contains the testing info
     message : str
         Message to report about the state
@@ -672,7 +670,7 @@ def check_perfect(info, results):
 
     Parameters
     ----------
-    info : dict
+    info : esm_tests.Info
         Dictionary with the general info about the tests.
     results : dict
         Dictionary containing the results of the testing.
@@ -754,477 +752,3 @@ def exist_files(files, path, version):
                 files_checked = False
 
     return files_checked
-
-
-#######################################################################################
-# UTILITIES
-#######################################################################################
-def get_rel_paths_compare_files(info, cfile, v, this_test_dir):
-    """
-    Returns the relative paths of the files in ``last_tested`` the corresponding ones
-    in the current experiment.
-
-    Parameters
-    ----------
-    info : dict
-        Dictionary that contains the testing info
-    cfile : str
-        Name of the current file
-    v : dict
-        Dictionary containing script variables such as ``version``, ``comp_command``, ...
-    this_test_dir : str
-        Path of the current experiment
-
-    Returns
-    -------
-    subpaths_source : list
-        Relative paths of the file in the current experiment
-    subpaths_target : list
-        Relative paths of the file in the ``last_tested`` folder
-    """
-    # Load relevant variables from ``info``
-    user_info = info["user"]
-    # Initialize ``subpaths`` list
-    subpaths = []
-    # If the file type is ``comp-``, append all the files that match that string to
-    # ``subpaths``
-    if cfile == "comp-":
-        for f in os.listdir(f"{user_info['test_dir']}/{this_test_dir}"):
-            if cfile in f:
-                subpaths.append(f"{this_test_dir}/{f}")
-        if len(subpaths) == 0:
-            logger.error("\t\tNo 'comp-*.sh' file found!")
-    # If the file type matches ``.run`` or ``finished_config``, append all the files
-    # that match that string to ``subpaths``. These files are taken always from the
-    # first ``run_`` folder so that a check test and an actual test files are
-    # comparable
-    elif cfile in [".run", "finished_config"]:
-        files_to_folders = {".run": "scripts", "finished_config": "config"}
-        ctype = files_to_folders[cfile]
-        ldir = os.listdir(f"{user_info['test_dir']}/{this_test_dir}")
-        ldir.sort()
-        for f in ldir:
-            # Take the first run directory
-            if "run_" in f:
-                cf_path = f"{this_test_dir}/{f}/{ctype}/"
-                cfiles = glob.glob(f"{user_info['test_dir']}/{cf_path}/*{cfile}*")
-                # If not found, try in the general directory
-                if len(cfiles) == 0:
-                    cf_path = f"{this_test_dir}/{ctype}/"
-                    cfiles = glob.glob(f"{user_info['test_dir']}/{cf_path}/*{cfile}*")
-                # Make sure we always take the first run
-                cfiles.sort()
-                num = 0
-                if len(cfiles)==0:
-                    break
-                if os.path.islink(
-                    f"{user_info['test_dir']}/{cf_path}/{cfiles[num].split('/')[-1]}"
-                ):
-                    num = 1
-                subpaths.append(f"{cf_path}/{cfiles[num].split('/')[-1]}")
-                break
-    elif cfile == "namelists":
-        # Get path of the finished_config
-        s_config_yaml, _ = get_rel_paths_compare_files(
-            info, "finished_config", v, this_test_dir
-        )
-        if len(s_config_yaml)>0:
-            namelists = extract_namelists(f"{user_info['test_dir']}/{s_config_yaml[0]}")
-            ldir = os.listdir(f"{user_info['test_dir']}/{this_test_dir}")
-            ldir.sort()
-            for f in ldir:
-                # Take the first run directory
-                if "run_" in f:
-                    cf_path = f"{this_test_dir}/{f}/work/"
-                    for n in namelists:
-                        if not os.path.isfile(f"{user_info['test_dir']}/{cf_path}/{n}"):
-                            logger.debug(f"'{cf_path}/{n}' does not exist!")
-                        subpaths.append(f"{cf_path}/{n}")
-                    break
-    else:
-        subpaths = [f"{this_test_dir}/{cfile}"]
-
-    # Remove run directory from the targets
-    subpaths_source = subpaths
-    subpaths_target = []
-    datestamp_format = re.compile(r"_[\d]{8}-[\d]{8}$")
-    for sp in subpaths:
-        sp_t = ""
-        pieces = sp.split("/")
-        for p in pieces:
-            if "run_" not in p:
-                sp_t += f"/{p}"
-        # Remove the datestamp
-        if datestamp_format.findall(sp_t):
-            sp_t = sp_t.replace(datestamp_format.findall(sp_t)[0], "")
-        subpaths_target.append(sp_t)
-
-    return subpaths_source, subpaths_target
-
-
-def extract_namelists(s_config_yaml):
-    """
-    Searches for the names of the namelists in the ``s_config_yaml`` (i.e.
-    ``*_finished_config.yaml``)
-
-    Parameters
-    ----------
-    s_config_yaml : str
-        Path to the ``*finished_config.yaml`` file.
-
-    Returns
-    -------
-    namelists : list
-        List of namelist names associated to this experiment.
-    """
-    # Read config file
-    with open(s_config_yaml, "r") as c:
-        config = yaml.load(c, Loader=yaml.FullLoader)
-
-    namelists = []
-    # Loop through the components to find the namelists
-    for component in config.keys():
-        namelists_component = config[component].get("namelists", [])
-        for nml in namelists_component:
-            config_sources = config[component].get("config_sources", {})
-            if nml in config_sources or any(
-                [nml in x for x in config_sources.values()]
-            ):
-                namelists.append(nml)
-
-    # Adds OASIS ``namcouple``
-    if "oasis3mct" in config:
-        namelists.append("namcouple")
-
-    return namelists
-
-
-#######################################################################################
-# OUTPUT
-#######################################################################################
-def print_diff(info, sscript, tscript, name, ignore_lines):
-    script_s = open(sscript).readlines()
-    script_t = open(tscript).readlines()
-
-    # Delete dictionaries to be ignored from the finished_config.yaml
-    if "finished_config.yaml" in tscript:
-        script_t = del_ignore_dicts(info, script_t)
-
-    # Substitute user lines in target string
-    script_t = clean_user_specific_info(info, script_t)
-
-    # Check for ignored lines
-    new_script_s = []
-    for line in script_s:
-        ignore_this = False
-        for iline in ignore_lines:
-            if re.search(iline, line):
-                ignore_this = True
-        if not ignore_this:
-            new_script_s.append(line)
-    script_s = new_script_s
-    new_script_t = []
-    for line in script_t:
-        ignore_this = False
-        for iline in ignore_lines:
-            if re.search(iline, line):
-                ignore_this = True
-        if not ignore_this:
-            new_script_t.append(line)
-    script_t = new_script_t
-
-    diffobj = difflib.SequenceMatcher(a=script_s, b=script_t)
-    differences = ""
-    if diffobj.ratio() == 1:
-        logger.info(f"\t\t'{name}' files are identical")
-        identical = True
-    else:
-        # Find differences
-        pdifferences = ""
-        for line in color_diff(difflib.unified_diff(script_s, script_t)):
-            differences += line
-            pdifferences += f"\t\t{line}"
-
-        logger.info(f"\n\tDifferences in \x1b[35m{name}:\x1b[0m\n{pdifferences}\n")
-        # input("Press enter to continue...")
-        identical = False
-
-    return identical, differences
-
-
-def del_ignore_dicts(info, yaml_file):
-    new_yaml_file = []
-    indentation_level_key = 0
-    dict_key_found = False
-    in_dict = False
-    for line in yaml_file:
-        indentation_level = len(line) - len(line.lstrip(" "))
-        if dict_key_found:
-            in_dict = indentation_level > indentation_level_key
-            in_dict |= (
-                indentation_level == indentation_level_key
-                and line.replace(" ", "")[0:1] == "- ")
-            if not in_dict:
-                dict_key_found = False
-        dict_key_in_line = False
-        for ig_key in info["ignore"]["finished_config_dicts"]:
-            dict_key_in_line |= bool(re.search(f"{ig_key}:", line))
-        if dict_key_in_line:
-            dict_key_found = True
-            indentation_level_key = indentation_level
-        elif in_dict:
-            pass
-        else:
-            new_yaml_file.append(line)
-    return new_yaml_file
-
-
-def save_files(info, user_choice):
-    # Load relevant variables from ``info``
-    scripts_info = info["scripts"]
-    actually_run = info["actually_run"]
-    user_info = info["user"]
-    last_tested_dir = info["last_tested_dir"]
-    runscripts_dir = get_runscripts_dir()
-    this_computer = info["this_computer"]
-    # ``user_choice`` is set through the ``--save`` flag. If the flag is not used the
-    # main program reaches this point and asks the user about saving files in
-    # ``last_tested``. If the flag is used alone or the value is set to ``True``, then
-    # it does save without asking. If the flag is used and the value is set to ``False``
-    # then it doesn't save and it doesn't ask
-    if not user_choice:
-        not_answered = True
-        while not_answered:
-            # Ask if saving of the compared files is required
-            save = input(
-                f"Would you like to save the files in the '{last_tested_dir}' folder for later "
-                + "comparisson and/or committing to GitHub?[y/n]: "
-            )
-            if save == "y":
-                not_answered = False
-            elif save == "n":
-                logger.info(f"No files will be saved in '{last_tested_dir}'")
-                not_answered = False
-                return
-            else:
-                print(f"'{save}' is not a valid answer!")
-
-    # Select test types
-    if info["actually_compile"]:
-        test_type_c = "actual"
-    else:
-        test_type_c = "check"
-    if actually_run:
-        test_type_r = "actual"
-    else:
-        test_type_r = "check"
-
-    logger.info(f"Saving files to '{last_tested_dir}'...")
-    # Loop through models
-    for model, scripts in scripts_info.items():
-        # Ignore ``general``
-        if model == "general":
-            continue
-        # Load the model's ``config.yaml``
-        model_config = f"{runscripts_dir}/{model}/config.yaml"
-        if not os.path.isfile(model_config):
-            logger.error(f"'{model_config}' not found!")
-        with open(model_config, "r") as c:
-            config_test = yaml.load(c, Loader=yaml.FullLoader)
-        # Get name patters from the files to be compared
-        compare_files_comp = copy.deepcopy(compare_files["comp"])
-        compare_files_comp.extend(
-            config_test.get("comp", {}).get(test_type_c, {}).get("compare", [])
-        )
-        compare_files_run = copy.deepcopy(compare_files["run"])
-        # TODO: The iterative coupling needs a rework. Therefore, no testing for files
-        # is develop. Include the tests after iterative coupling is reworked
-        if next(iter(scripts.values()))["iterative_coupling"]:
-            compare_files_run = []
-        compare_files_run.extend(
-            config_test.get("run", {}).get(test_type_r, {}).get("compare", [])
-        )
-        # Loop through scripts
-        for script, v in scripts.items():
-            version = v["version"]
-            # Loop through comp and run
-            for mode in ["comp", "run"]:
-                if mode == "comp":
-                    this_compare_files = compare_files_comp
-                    subfolder = f"{model}-{version}"
-                elif mode == "run":
-                    this_compare_files = compare_files_run
-                    subfolder = f"{script}"
-                    save_exp_date(info, model, script)
-                this_test_dir = f"{mode}/{model}/{subfolder}/"
-                # Loop through comparefiles
-                for cfile in this_compare_files:
-                    subpaths_source, subpaths_target = get_rel_paths_compare_files(
-                        info, cfile, v, this_test_dir
-                    )
-                    for sp, sp_t in zip(subpaths_source, subpaths_target):
-                        target_path = f"{last_tested_dir}/{this_computer}/{sp_t}"
-                        if os.path.isfile(target_path):
-                            logger.debug(
-                                f"\t'{sp_t}' file in '{last_tested_dir}/{this_computer}' will be overwritten"
-                            )
-                        if not os.path.isdir(os.path.dirname(target_path)):
-                            os.makedirs(os.path.dirname(target_path))
-                        shutil.copy2(f"{user_info['test_dir']}/{sp}", target_path)
-
-                        # Modify lines containing user-specific information
-                        with open(target_path) as f:
-                            stext = clean_user_specific_info(info, f.read())
-                        with open(target_path, "w") as f:
-                            f.write(stext)
-
-                        # If check run and file is the ``finished_config.yaml``
-                        if "finished_config.yaml" in target_path:
-                            with open(target_path) as f:
-                                yaml_file = f.readlines()
-                            # Delete dictionaries to be ignored
-                            yaml_file = del_ignore_dicts(info, yaml_file)
-                            with open(target_path, "w") as f:
-                                f.write("".join(yaml_file))
-
-    if info["actually_compile"] and info["actually_run"]:
-        # Load current state
-        with open(get_state_yaml_path(), "r") as st:
-            current_state = yaml.load(st, Loader=yaml.FullLoader)
-        # Update with this results
-        results = format_results(info)
-        current_state = deep_update(current_state, results)
-        current_state = sort_dict(current_state)
-        with open(get_state_yaml_path(), "w") as st:
-            state = yaml.dump(current_state)
-            st.write(state)
-
-
-def save_exp_date(info, model, script):
-    user_info = info["user"]
-    if info["actually_run"]:
-        exp_time = os.path.getmtime(f"{user_info['test_dir']}/run/{model}/{script}")
-        exp_human_time = time.strftime("%d-%m-%Y", time.gmtime(exp_time))
-        this_script = info["scripts"][model][script]
-        success = True
-        #if all(this_script["state"].values()):
-        this_script["date"] = exp_human_time
-
-
-def print_results(results, info):
-    colorama.init(autoreset=True)
-
-    if info.get("bulletpoints", False):
-        bp = "- "
-    else:
-        bp = ""
-
-    if info["in_github"]:
-        first_tab = "|   "
-    else:
-        first_tab = "    "
-
-    logger.info("")
-    logger.info("")
-    logger.info(f"{bs}RESULTS{be}")
-    logger.info("")
-    for model, versions in results.items():
-        logger.info(f"{bp}{colorama.Fore.CYAN}{model}:")
-        for version, scripts in versions.items():
-            logger.info(f"{first_tab}{bp}{colorama.Fore.MAGENTA}{version}:")
-            for script, computers in scripts.items():
-                logger.info(f"{first_tab}    {bp}{colorama.Fore.WHITE}{script}:")
-                for computer, data in computers.items():
-                    text = data["compilation"]
-                    if "compiles" == text:
-                        text_color = colorama.Fore.GREEN
-                    elif "identical" in text and not info["actually_compile"]:
-                        text_color = colorama.Fore.GREEN
-                    elif "differ" in text:
-                        text_color = colorama.Fore.YELLOW
-                    else:
-                        text_color = colorama.Fore.RED
-                    compilation = f"{text_color}{text}"
-                    text = data["run"]
-                    if "runs" == text:
-                        text_color = colorama.Fore.GREEN
-                    elif "identical" in text and not info["actually_run"]:
-                        text_color = colorama.Fore.GREEN
-                    elif "differ" in text:
-                        text_color = colorama.Fore.YELLOW
-                    else:
-                        text_color = colorama.Fore.RED
-                    run = f"{text_color}{text}"
-                    if data.get("date"):
-                        text_color = colorama.Fore.WHITE
-                        date = f"\t{text_color}{data['date']}"
-                    else:
-                        date = ""
-                    logger.info(
-                        f"{first_tab}        {bp}{colorama.Fore.WHITE}{computer}:\t{compilation}\t{run}{date}"
-                    )
-    logger.info(f"{colorama.Fore.WHITE}")
-    logger.info("")
-
-
-def format_results(info):
-    # Load current state
-    with open(get_state_yaml_path(), "r") as st:
-        current_state = yaml.load(st, Loader=yaml.FullLoader)
-
-    scripts_info = info["scripts"]
-    results = {}
-    for model, scripts in scripts_info.items():
-        if model == "general":
-            continue
-        results[model] = {}
-        for script, v in scripts.items():
-            version = v["version"]
-            results[model][version] = results[model].get(version, {})
-            results[model][version][script] = results[model][version].get(script, {})
-            state = v["state"]
-
-            if info["actually_compile"]:
-                compilation = "compiles"
-            else:
-                compilation = "comp files identical"
-
-            if not state.get("comp_files_identical", True):
-                compilation = "comp files differ"
-            if not state.get("comp_files", True):
-                compilation = "comp files not found"
-            if not state["comp"]:
-                compilation = "compilation failed"
-
-            if info["actually_run"]:
-                run = "runs"
-            else:
-                run = "run files identical"
-
-            if not state.get("submission_files_identical", True):
-                run = "run files differ"
-            if not state["submission"]:
-                run = "submission failed"
-            if not state.get("run_files", True):
-                run = "run files not found"
-            if not state.get("run_finished", True):
-                run = "run failed"
-
-            date = info["scripts"][model][script].get("date")
-
-            results[model][version][script][info["this_computer"]] = {
-                "compilation": compilation,
-                "run": run,
-                "date": date,
-            }
-
-    return results
-
-
-def sort_dict(dict_to_sort):
-    if isinstance(dict_to_sort, dict):
-        dict_to_sort = {key: dict_to_sort[key] for key in sorted(dict_to_sort.keys())}
-        for key, value in dict_to_sort.items():
-            dict_to_sort[key] = sort_dict(value)
-
-    return dict_to_sort
